Move linked list implementation to `types` directory
