+++ core/types/list.cpp
